Massive change to the file structure. and coding layout/style
I am almost starting over. This time around I am taking into account
all of the preprocessor commands. It handles all of the version
checking for windows. But it also gives the user some definable options.
things like setting the available API based on the Windows type
server, rt, desktop, etc...

I also built a idl parser. this is a much cleaner mechanism to handle
most com interfaces. Not all interfaces are located in idl files tho.

I also built a database of sorts that holds all of the windows core
DLL file names coupled with all of the available functions in that are
in a specific dll file. This makes the work of parsing the function
calls from the h files a whole lot easier. I only currently have a
database for windows 7. I am thinking I am going to need to install
Windows 10 into a VM and iterate over all of the dll files. making the
calls to dll walker using subprocess to obtain the function names.
Once that is done the hunt based on dll priority begins. to filter out
the duplicate functions.

This whole new mechanism seems to be alot better in terms of parsing
the h files. There is a huge amount of reduction in the manual work
needed to clean up the errors. The new builder I have designed in a way
that it will be able to folllow the include path and create files
along the way.

I still need to make a mechanism that will build the __all__. tho at
the moment I find it very easy to simply paste some code at the bottom
that reads the modules __dict__ and spits out a nice formatted __all__.
This might be the best way since there is still manual work that needs
to be done to each file. If I could get the conversions to be 100% then
i could simply run the built code using exec passing in a namespace.
I do not ever see the conversion process 100%.