Adds more to wdm_h and also changes the converter
The code converter has been changed so that all libraries, interfaces,
co classes, structures and unions are forward declared. this removes
all kinds of complexities with placing specific pieces in ine proper
locations inside the file. This is something that is not that much of
an issue with smaller files. But when working woth a 40K line file
It can become a real mess very quickly.