Updates the code_converter package
There was an issue with a multi line preprocesssor statement having
part of the statement chopped off. this is now fixed.

when the _fields_ gets declared in a structure or union any comments
were added to the end of the field. Because Python has indent syntax
and C/C++ does not we would bypass the PEP 79 character width. it also
made it a general annoyance when trying to read the comments. This issue
does not exist in C/C++ code because in most cases the lines are not
indented, specifically when dealing with the preprocessor stuff.

So i moved the comments so they print out before the field item and I also
split the comment down to fit the space adding lines as necessary. I have
the split take place on the space closest to the 79 characters unless
the split would occur inside of a code block typically inside of
[], (), <>, {}
if this occurs it will find the next available space after the
brackets are closed